requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.2.13
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
beautifulsoup4>=4.9.3

